+++ src/platform/web.rs
@@ -35,7 +35,12 @@ use crate::platform_impl::PlatformCustomCursorBuilder;
 use crate::window::CustomCursor;
 use crate::window::{Window, WindowBuilder};
 
+#[cfg(wasm_platform)]
 use web_sys::HtmlCanvasElement;
+
+#[cfg(not(wasm_platform))]
+#[doc(hidden)]
+pub struct HtmlCanvasElement(());
 
 pub trait WindowExtWebSys {
     /// Only returns the canvas if called from inside the window context (the